Split OpenAI Agents sandbox providers
